backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyephem==9.99
Pillow==10.2.0
pydantic==2.5.3
python-multipart==0.0.6
httpx==0.26.0
